requirements.txt
pyyaml
shiny @ git+https://github.com/posit-dev/py-shiny@markdown-stream-context
python-dotenv
chatlas[openai]
